+++ client/src/components/LoginImageCarousel.jsx
