+++ client/src/pages/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import NavBar from "../components/NavBarDashboard";
+import NavBar from "../components/DashboardNavBar";
 import DashboardHub from '../components/DashboardHub';
 
 const Dashboard = () => {
